Rename: Rename LoginResult to LoginEntity
Rename domain/model package to domain/entity/type
