+++ core/domain/src/main/java/kr/co/domain/entity/LoginEntity.kt
@@ -1,0 +1,8 @@
+package kr.co.domain.entity
+
+import kr.co.domain.entity.type.AuthType
+
+data class LoginEntity(
+    val type: AuthType,
+    val token: String,
+)
