--- core/domain/src/main/java/kr/co/domain/model/LoginResult.kt
@@ -1,6 +1,0 @@
-package kr.co.domain.model
-
-data class LoginResult(
-    val type: AuthType,
-    val token: String,
-)
