+++ core/domain/src/main/java/kr/co/domain/proivder/SocialLoginProvider.kt
@@ -1,10 +1,10 @@
 package kr.co.domain.proivder
 
-import kr.co.domain.model.AuthType
-import kr.co.domain.model.LoginResult
+import kr.co.domain.entity.type.AuthType
+import kr.co.domain.entity.LoginEntity
 
 interface SocialLoginProvider {
-    suspend fun login(type: AuthType): LoginResult
+    suspend fun login(type: AuthType): LoginEntity
 
     suspend fun logout(type: AuthType)
 }
