+++ core/domain/src/main/java/kr/co/domain/repository/AuthRepository.kt
@@ -1,6 +1,6 @@
 package kr.co.domain.repository
 
-import kr.co.domain.model.AuthType
+import kr.co.domain.entity.type.AuthType
 
 interface AuthRepository {
