+++ core/oauth/src/main/kotlin/kr/co/oauth/GoogleLoginProviderImpl.kt
@@ -1,18 +1,14 @@
 package kr.co.oauth
 
-import android.content.Context
-import dagger.hilt.android.qualifiers.ActivityContext
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kr.co.domain.model.AuthType
-import kr.co.domain.model.LoginResult
+import kr.co.domain.entity.type.AuthType
+import kr.co.domain.entity.LoginEntity
 import kr.co.domain.proivder.SocialLoginProvider
 import javax.inject.Inject
-import javax.inject.Named
 
 internal class GoogleLoginProviderImpl @Inject constructor(
 ) : SocialLoginProvider {
-    override suspend fun login(type: AuthType): LoginResult {
-        return LoginResult(AuthType.GOOGLE,"")
+    override suspend fun login(type: AuthType): LoginEntity {
+        return LoginEntity(AuthType.GOOGLE,"")
     }
 
     override suspend fun logout(type: AuthType) {
