+++ core/oauth/src/main/kotlin/kr/co/oauth/di/OAuthModule.kt
@@ -2,20 +2,16 @@ package kr.co.oauth.di
 
 import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityComponent
-import dagger.hilt.android.scopes.ActivityRetainedScoped
 import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.IntoMap
-import kr.co.domain.model.AuthType
+import kr.co.domain.entity.type.AuthType
 import kr.co.domain.proivder.SocialLoginProvider
 import kr.co.oauth.GoogleLoginProviderImpl
 import kr.co.oauth.KakaoLoginProviderImpl
 import kr.co.oauth.NaverLoginProviderImpl
 import kr.co.oauth.SocialLoginProviderImpl
 import kr.co.oauth.annotation.SocialLoginKey
-import javax.inject.Provider
 import javax.inject.Singleton
 
 @Module
